Move FilesEntitlements validation to a separate class
