+++ libs/entitlement/src/main/java/org/elasticsearch/entitlement/initialization/EntitlementInitialization.java
@@ -11,7 +11,6 @@ package org.elasticsearch.entitlement.initialization;
 
 import org.elasticsearch.core.Booleans;
 import org.elasticsearch.core.PathUtils;
-import org.elasticsearch.core.Strings;
 import org.elasticsearch.core.internal.provider.ProviderLocator;
 import org.elasticsearch.entitlement.bootstrap.EntitlementBootstrap;
 import org.elasticsearch.entitlement.bridge.EntitlementChecker;
@@ -21,7 +20,6 @@ import org.elasticsearch.entitlement.instrumentation.Instrumenter;
 import org.elasticsearch.entitlement.instrumentation.MethodKey;
 import org.elasticsearch.entitlement.instrumentation.Transformer;
 import org.elasticsearch.entitlement.runtime.api.ElasticsearchEntitlementChecker;
-import org.elasticsearch.entitlement.runtime.policy.FileAccessTree;
 import org.elasticsearch.entitlement.runtime.policy.PathLookup;
 import org.elasticsearch.entitlement.runtime.policy.Policy;
 import org.elasticsearch.entitlement.runtime.policy.PolicyManager;
@@ -60,7 +58,6 @@ import java.nio.file.spi.FileSystemProvider;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -345,7 +342,7 @@ public class EntitlementInitialization {
             )
         );
 
-        validateFilesEntitlements(pluginPolicies, pathLookup);
+        FilesEntitlementsValidation.validate(pluginPolicies, pathLookup);
 
         return new PolicyManager(
             serverPolicy,
@@ -357,74 +354,6 @@ public class EntitlementInitialization {
             pathLookup,
             bootstrapArgs.suppressFailureLogClasses()
         );
-    }
-
-    // package visible for tests
-    static void validateFilesEntitlements(Map<String, Policy> pluginPolicies, PathLookup pathLookup) {
-        Set<Path> readAccessForbidden = new HashSet<>();
-        pathLookup.getBaseDirPaths(PLUGINS).forEach(p -> readAccessForbidden.add(p.toAbsolutePath().normalize()));
-        pathLookup.getBaseDirPaths(MODULES).forEach(p -> readAccessForbidden.add(p.toAbsolutePath().normalize()));
-        pathLookup.getBaseDirPaths(LIB).forEach(p -> readAccessForbidden.add(p.toAbsolutePath().normalize()));
-        Set<Path> writeAccessForbidden = new HashSet<>();
-        pathLookup.getBaseDirPaths(CONFIG).forEach(p -> writeAccessForbidden.add(p.toAbsolutePath().normalize()));
-        for (var pluginPolicy : pluginPolicies.entrySet()) {
-            for (var scope : pluginPolicy.getValue().scopes()) {
-                var filesEntitlement = scope.entitlements()
-                    .stream()
-                    .filter(x -> x instanceof FilesEntitlement)
-                    .map(x -> ((FilesEntitlement) x))
-                    .findFirst();
-                if (filesEntitlement.isPresent()) {
-                    var fileAccessTree = FileAccessTree.withoutExclusivePaths(filesEntitlement.get(), pathLookup, null);
-                    validateReadFilesEntitlements(pluginPolicy.getKey(), scope.moduleName(), fileAccessTree, readAccessForbidden);
-                    validateWriteFilesEntitlements(pluginPolicy.getKey(), scope.moduleName(), fileAccessTree, writeAccessForbidden);
-                }
-            }
-        }
-    }
-
-    private static IllegalArgumentException buildValidationException(
-        String componentName,
-        String moduleName,
-        Path forbiddenPath,
-        FilesEntitlement.Mode mode
-    ) {
-        return new IllegalArgumentException(
-            Strings.format(
-                "policy for module [%s] in [%s] has an invalid file entitlement. Any path under [%s] is forbidden for mode [%s].",
-                moduleName,
-                componentName,
-                forbiddenPath,
-                mode
-            )
-        );
-    }
-
-    private static void validateReadFilesEntitlements(
-        String componentName,
-        String moduleName,
-        FileAccessTree fileAccessTree,
-        Set<Path> readForbiddenPaths
-    ) {
-
-        for (Path forbiddenPath : readForbiddenPaths) {
-            if (fileAccessTree.canRead(forbiddenPath)) {
-                throw buildValidationException(componentName, moduleName, forbiddenPath, READ);
-            }
-        }
-    }
-
-    private static void validateWriteFilesEntitlements(
-        String componentName,
-        String moduleName,
-        FileAccessTree fileAccessTree,
-        Set<Path> writeForbiddenPaths
-    ) {
-        for (Path forbiddenPath : writeForbiddenPaths) {
-            if (fileAccessTree.canWrite(forbiddenPath)) {
-                throw buildValidationException(componentName, moduleName, forbiddenPath, READ_WRITE);
-            }
-        }
     }
 
     private static Path getUserHome() {
